+++ src/newtonRaphsonIteration/newtonRaphsonIteration.cpp
@@ -10,15 +10,15 @@ using namespace std;
 #include "fxCompute.hpp"
 #include "differentiate.hpp"
 
-int newtonRaphsonIteration(vector<vector<string>> polynomial, int xn)
+double newtonRaphsonIteration(vector<vector<string>> polynomial, double xn)
 {
 
     vector<vector<string>> DerivativeFx = differentiate(polynomial);
 
-    int FXn = computeFX(polynomial, xn);
-    int DydxXn = computeFX(DerivativeFx, xn);
+    double FXn = computeFX(polynomial, xn);
+    double DydxXn = computeFX(DerivativeFx, xn);
 
-    int XnPlus1 = xn - (FXn / DydxXn);
+    double XnPlus1 = xn - (FXn / DydxXn);
 
     return XnPlus1;
 }
